Move llama modules to generic modules dir and create builder
